Discord BOT/requirements.txt
discord.py
requests
translate